feat: ammonium and specific conducatance
This is a big commit. The main updates are having added queries for
analysis-ready ammonium and specific conductance data in addition to
corresponding data summaries.

The document has been reworked such that the queries (and the outline)
reflect generating views, temporary tables, or exports.

The query to genrate ranges was refactored to accommodate occurrences
where more than one fire occurred on a given day in a given catchment.

Querying nitrate data was fixed so that both values and units were
standarsized.

The call to ST_Area was refactored such that TRUE was passed to the
second argument of the function, thus eliminating the need to cast the
geometry to a geography.